Disable braindead KDE3/4 prefix detection.
The current detection of the KDE 3 and KDE 4 prefixes is done running
kde-config and kde4-config, resp.
This can be problematic, as these two could create files in ~/.kde,
thus breaking when the home of the current user (eg in buildds) is read-only.
Given the build system sets anyway "/usr" as prefix of KDE 3 and KDE 4
if they were not detected, then just let it to.

Bug-Debian: http://bugs.debian.org/541667
Author: Pino Toscano <pino@kde.org>